controller

+++ src/middlewares/validates/sessionAuth.middleware.ts
@@ -5,25 +5,22 @@ import {StatusId} from "../../constants/status";
 import {sessionAuthTTL} from "../../config/session/session.auth";
 import logMiddleware from "../log.middleware";
 
-const check = async (
-    req: FastifyRequest,
-    res: FastifyReply
-) => {
+const check = async (req: FastifyRequest,res: FastifyReply) => {
     await logMiddleware.error(req, res, async () => {
         let serviceResult = new Result();
 
-        if (req.sessionAuth && req.sessionAuth.data) {
+        if (req.sessionAuth && req.sessionAuth.data() && req.sessionAuth.user) {
             if (req.sessionAuth.user?.ip != req.ip) {
                 await new Promise(resolve => {
                     req.sessionAuth.delete();
-                    resolve();
-                })
+                    resolve(1);
+                });
             }
         }
 
         if (
             (typeof req.sessionAuth === "undefined" || typeof req.sessionAuth.user === "undefined") ||
-            !(await userService.getOne({_id: req.sessionAuth.user?._id, statusId: StatusId.Active}))
+            !(await userService.getOne({_id: req.sessionAuth.user.userId, statusId: StatusId.Active}))
         ) {
             serviceResult.status = false;
             serviceResult.errorCode = ErrorCodes.notLoggedIn;
@@ -38,21 +35,20 @@ const check = async (
     });
 };
 
-const reload = async (
-    req: FastifyRequest,
-    res: FastifyReply
-) => {
+const reload = async (req: FastifyRequest,res: FastifyReply) => {
     await logMiddleware.error(req, res, async () => {
-        if (req.sessionAuth && req.sessionAuth.data) {
-            if (Number(new Date().diffSeconds(new Date(req.sessionAuth.user?.updatedAt ?? ""))) > sessionAuthTTL) {
+        if (req.sessionAuth && req.sessionAuth.data() && req.sessionAuth.user) {
+            if (Number(new Date().diffSeconds(new Date(req.sessionAuth.user.updatedAt ?? ""))) > sessionAuthTTL) {
                 await new Promise(resolve => {
                     req.sessionAuth.delete();
-                    resolve();
+                    resolve(1);
                 })
             }
         }
-        if (req.sessionAuth && req.sessionAuth.data) {
-            req.sessionAuth.data.updatedAt = Date.now();
+        if (req.sessionAuth && req.sessionAuth.data() && req.sessionAuth.user) {
+            req.sessionAuth.set("user", {
+                ...req.sessionAuth.user
+            });
         }
     });
 };
