+++ src/routers/routes/post.route.ts
@@ -12,11 +12,11 @@ export default function (fastify: FastifyInstance, opts: any, done: () => void) 
     fastify.get(PostEndPoint.GET, { preHandler: [requestMiddleware.check(postSchema.getMany)] }, postController.getMany);
     fastify.get(PostEndPoint.GET_COUNT, { preHandler: [requestMiddleware.check(postSchema.getCount)] }, postController.getCount);
     fastify.get(PostEndPoint.GET_WITH_ID, { preHandler: [requestMiddleware.check(postSchema.get)] }, postController.getOne);
-    fastify.post(PostEndPoint.ADD, { preHandler: [requestMiddleware.check(postSchema.post), sessionMiddleware.check, (req) => permissionMiddleware.check(permissionUtil.getPostPermission(req)), postMiddleware.checkUrl] }, postController.add);
-    fastify.put(PostEndPoint.UPDATE_VIEW_WITH_ID, { preHandler: [requestMiddleware.check(postSchema.putOneView), viewMiddleware.checkOne, postMiddleware.checkOne] }, postController.updateOneView);
-    fastify.put(PostEndPoint.UPDATE_RANK_WITH_ID, { preHandler: [requestMiddleware.check(postSchema.putOneRank), sessionMiddleware.check, permissionMiddleware.check, postMiddleware.checkOne] }, postController.updateOneRank);
-    fastify.put(PostEndPoint.UPDATE_STATUS, { preHandler: [requestMiddleware.check(postSchema.putManyStatus), sessionMiddleware.check, permissionMiddleware.check, postMiddleware.checkMany] }, postController.updateManyStatus);
-    fastify.put(PostEndPoint.UPDATE_WITH_ID, { preHandler: [requestMiddleware.check(postSchema.put), sessionMiddleware.check, permissionMiddleware.check, postMiddleware.checkOne, postMiddleware.checkUrl] }, postController.updateOne);
-    fastify.delete(PostEndPoint.DELETE, { preHandler: [requestMiddleware.check(postSchema.deleteMany), sessionMiddleware.check, permissionMiddleware.check, postMiddleware.checkMany] }, postController.deleteMany);
+    fastify.post(PostEndPoint.ADD, { preHandler: [requestMiddleware.check(postSchema.post), sessionMiddleware.check, permissionMiddleware.check(permissionUtil.getPostPermission)] }, postController.add);
+    fastify.put(PostEndPoint.UPDATE_VIEW_WITH_ID, { preHandler: [requestMiddleware.check(postSchema.putView), viewMiddleware.check, postMiddleware.check] }, postController.updateOneView);
+    fastify.put(PostEndPoint.UPDATE_RANK_WITH_ID, { preHandler: [requestMiddleware.check(postSchema.putRank), sessionMiddleware.check, permissionMiddleware.check(permissionUtil.getPostPermission), postMiddleware.check] }, postController.updateOneRank);
+    fastify.put(PostEndPoint.UPDATE_STATUS, { preHandler: [requestMiddleware.check(postSchema.putManyStatus), sessionMiddleware.check, permissionMiddleware.check(permissionUtil.getPostPermission), postMiddleware.checkMany] }, postController.updateManyStatus);
+    fastify.put(PostEndPoint.UPDATE_WITH_ID, { preHandler: [requestMiddleware.check(postSchema.put), sessionMiddleware.check, permissionMiddleware.check(permissionUtil.getPostPermission), postMiddleware.check] }, postController.updateOne);
+    fastify.delete(PostEndPoint.DELETE, { preHandler: [requestMiddleware.check(postSchema.deleteMany), sessionMiddleware.check, permissionMiddleware.check(permissionUtil.getPostPermission), postMiddleware.checkMany] }, postController.deleteMany);
     done();
 }
